* sbitmap.h: Make SBITMAP_ELT_BITS unsigned.
From-SVN: r31912

+++ gcc/sbitmap.h
@@ -25,7 +25,7 @@ Boston, MA 02111-1307, USA.  */
    It should be straightforward to convert so for now we keep things simple
    while more important issues are dealt with.  */
 
-#define SBITMAP_ELT_BITS HOST_BITS_PER_WIDE_INT
+#define SBITMAP_ELT_BITS ((unsigned) HOST_BITS_PER_WIDE_INT)
 #define SBITMAP_ELT_TYPE unsigned HOST_WIDE_INT
 
 typedef struct simple_bitmap_def {
